[UPDATE]{module structure} module directories restructured
