+++ src/components/mainComponents/Manual.js
