+++ src/constants/credentials.js
@@ -1,0 +1,3 @@
+export const username = "064c5a84d81a49048660883f0b4fee52";
+export const password = "921f1b7184164a8798d7d43b1adef4bc";
+export const GAID = "UA-90376709-3";
